Rename TypedPAutomaton to PAutomaton
