+++ src/include/pdaaal/SolverInstance.h
@@ -27,7 +27,7 @@
 #ifndef PDAAAL_SOLVERINSTANCE_H
 #define PDAAAL_SOLVERINSTANCE_H
 
-#include "TypedPAutomaton.h"
+#include "PAutomaton.h"
 #include "pdaaal/cegar/AbstractionPDA.h"
 #include "pdaaal/cegar/AbstractionPAutomaton.h"
 #include "PAutomatonProduct.h"
@@ -42,7 +42,7 @@ namespace pdaaal {
     class SolverInstance {
     public:
         using pda_t = PDA<T,W,fut::type::vector,state_t,skip_state_mapping>;
-        using pautomaton_t = TypedPAutomaton<T,W,state_t,skip_state_mapping, trace_info_type>;
+        using pautomaton_t = PAutomaton<T,W,state_t,skip_state_mapping, trace_info_type>;
         using product_t = PAutomatonProduct<pda_t, pautomaton_t, W, trace_info_type>;
         SolverInstance(pda_t&& pda,
                        const NFA<T>& initial_nfa, const std::vector<size_t>& initial_states,
@@ -81,8 +81,8 @@ namespace pdaaal {
                    const NFA<label_t>& final_nfa, const std::vector<size_t>& final_states) -> SolverInstance<label_t,W,state_t,skip_state_mapping,trace_info_type>;
     template<typename label_t, typename W, typename state_t, bool skip_state_mapping, TraceInfoType trace_info_type>
     SolverInstance(PDA<label_t,W,fut::type::vector,state_t,skip_state_mapping>&& pda,
-                   TypedPAutomaton<label_t,W,state_t,skip_state_mapping,trace_info_type> initial,
-                   TypedPAutomaton<label_t,W,state_t,skip_state_mapping,trace_info_type> final) -> SolverInstance<label_t,W,state_t,skip_state_mapping,trace_info_type>;
+                   PAutomaton<label_t,W,state_t,skip_state_mapping,trace_info_type> initial,
+                   PAutomaton<label_t,W,state_t,skip_state_mapping,trace_info_type> final) -> SolverInstance<label_t,W,state_t,skip_state_mapping,trace_info_type>;
 
     template <typename T, typename W>
     class AbstractionSolverInstance {
