+++ src/include/pdaaal/internal/PAutomaton.h
@@ -24,8 +24,8 @@
  * Created on 08-01-2020.
  */
 
-#ifndef PDAAAL_PAUTOMATON_H
-#define PDAAAL_PAUTOMATON_H
+#ifndef PDAAAL_INTERNAL_PAUTOMATON_H
+#define PDAAAL_INTERNAL_PAUTOMATON_H
 
 #include "PDA.h"
 #include "pdaaal/utils/fut_set.h"
@@ -777,4 +777,4 @@ namespace pdaaal::internal {
 
 }
 
-#endif //PDAAAL_PAUTOMATON_H
+#endif //PDAAAL_INTERNAL_PAUTOMATON_H
